cleaned and made distribution abc
